New WebRTC signaling sample WIP
Adding the initial code for the WebRTC signaling server sample. This is
based on the text chat server that uses WebSockets, but adds support
for inviting a user into two-way video chat. The signaling support code
has been added to the chatserver.js file, and early changes made to the
HTML and CSS, but signaling doesn’t yet take place. I just wanted to
get the code checked in for tracking work going forward.