Checkpoint: more client side work, tweaks server side
This is not functional at the moment. I am in the process of finishing
the design of the messages used to extend and accept chat invitations.
They’re simple but I want to get them right — part of the code in place
for handling these server-side was done before I had my epiphany and
understood really how signaling works, so it needs some changes. But I
promised to commit tonight, so here it is. :)