Capturing ongoing work on WebRTC sample
This code doesn’t work yet but I need to capture it so I can migrate it
to my laptop. Most of the code is in place for handling setting up the
video call now, but I’m still working through making sure all the logic
makes sense and adding in tracking log output to monitor it during
testing.