More for bug 1275883
Removes the .htaccess file in favor of adding script to the update.sh
script which revises the main httpd.conf to disable index listings and
to make all files and directories starting with “.git” result in 404
errors.

This is a much more reliable solution than what I was doing before, I
think.